实现Composer的支持
