+++ Example/index.php
@@ -21,6 +21,8 @@ define('APP_DEBUG',True);
 define('APP_PATH','./Application/');
 
 // 引入ThinkPHP入口文件
+
+require '../vendor/autoload.php';
 require './ThinkPHP/ThinkPHP.php';
 
 // 亲^_^ 后面不需要任何代码了 就是如此简单
