+++ composer.json
@@ -1,6 +1,6 @@
 {
     "name": "zt/unit-test",
-    "version":1.1,
+    "version":"1.2",
     "type": "library",
     "description": "Thinkphp Unit Test",
     "keywords": ["ThinkPHP","UnitTest"],
@@ -17,8 +17,6 @@
         "php": ">=5.3.0"
     },
     "autoload": {
-        "psr-4": {
-            "ZT\\Unit\\":"src"
-        }
+        "psr-4": {"":"src"}
     }
 }
